Refactor(cookie-log): react query 로직 분리
