+++ src/api/services/answer/record-paging/index.ts
@@ -1,13 +1,8 @@
-import { useSuspenseInfiniteQuery } from '@tanstack/react-query'
-
 import { authorizationInstance } from '@/api/instance'
-import {
-  API_ERROR_MESSAGES,
-  DATA_ERROR_MESSAGES,
-} from '@/constants/error-message'
+import { API_ERROR_MESSAGES } from '@/constants/error-message'
 import { AnswerRecord, Paging } from '@/types'
 
-type AnswerRecordPagingRequestParams = {
+export type AnswerRecordPagingRequestParams = {
   size?: number
   page?: string
   sort?: string[]
@@ -17,7 +12,7 @@ type AnswerRecordPagingResponse = {
   content: AnswerRecord[]
 } & Paging
 
-const getAnswerRecordPaging = async (
+export const getAnswerRecordPaging = async (
   params: AnswerRecordPagingRequestParams
 ) => {
   try {
@@ -37,45 +32,6 @@ const getAnswerRecordPaging = async (
     }
   } catch (error) {
     throw new Error(API_ERROR_MESSAGES.UNKNOWN_ERROR)
-  }
-}
-
-interface AnswerRecordPagingProps extends AnswerRecordPagingRequestParams {
-  initPageToken?: string
-}
-
-export const useAnswerRecordPaging = ({
-  size = 10,
-  sort,
-  initPageToken,
-}: AnswerRecordPagingProps) => {
-  const {
-    data,
-    status,
-    error,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-  } = useSuspenseInfiniteQuery({
-    queryKey: ['answer', 'record', initPageToken],
-    queryFn: ({ pageParam = initPageToken }) =>
-      getAnswerRecordPaging({ page: pageParam, size, sort }),
-    initialPageParam: initPageToken,
-    getNextPageParam: (lastPage) => lastPage.nextPageToken,
-  })
-
-  const answerRecords = data?.pages.flatMap((page) => page.records)
-
-  if (!answerRecords.length)
-    throw new Error(DATA_ERROR_MESSAGES.ANSWER_RECORD_NOT_FOUND)
-
-  return {
-    answerRecords,
-    status,
-    error,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
   }
 }
